# Backend/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResponseType__hostIds-string-Array__": {
				"properties": {
					"data": {
						"properties": {
							"hostIds": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						},
						"required": [
							"hostIds"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"ECreateHostBody": {
				"properties": {
					"hostCount": {
						"type": "number",
						"format": "double"
					},
					"gameId": {
						"type": "string"
					},
					"groupId": {
						"type": "string"
					}
				},
				"required": [
					"hostCount",
					"gameId",
					"groupId"
				],
				"type": "object"
			},
			"EHostIds": {
				"properties": {
					"hostIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"hostIds"
				],
				"type": "object"
			},
			"HostState": {
				"description": "Trạng thái host",
				"enum": [
					"lobby",
					"in-game",
					"ended"
				],
				"type": "string"
			},
			"GameMode": {
				"description": "Thể loại game",
				"enum": [
					"cafe"
				],
				"type": "string"
			},
			"HostLeaderboardItem": {
				"properties": {
					"avatar": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"playerId": {
						"type": "string"
					}
				},
				"required": [
					"score",
					"playerId"
				],
				"type": "object"
			},
			"ActivityBoardItem": {
				"properties": {
					"activity": {},
					"avatar": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"playerId": {
						"type": "string"
					}
				},
				"required": [
					"activity",
					"avatar",
					"username",
					"playerId"
				],
				"type": "object"
			},
			"Question": {
				"properties": {
					"correctAnswerId": {
						"type": "string"
					},
					"answers": {
						"items": {
							"properties": {
								"text": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"text",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"text": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"correctAnswerId",
					"answers",
					"text",
					"id"
				],
				"type": "object"
			},
			"Pick_HostInfo.Exclude_keyofHostInfo.personalResult-or-userInfo__": {
				"properties": {
					"hostId": {
						"type": "string"
					},
					"gameId": {
						"type": "string"
					},
					"gameSettings": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"description": "Cài đặt game"
					},
					"state": {
						"$ref": "#/components/schemas/HostState"
					},
					"gameMode": {
						"$ref": "#/components/schemas/GameMode"
					},
					"finalStandings": {
						"items": {
							"$ref": "#/components/schemas/HostLeaderboardItem"
						},
						"type": "array",
						"description": "Bảng xếp hạng"
					},
					"activitiesBoard": {
						"items": {
							"$ref": "#/components/schemas/ActivityBoardItem"
						},
						"type": "array",
						"description": "Lịch sử hoạt động"
					},
					"groupId": {
						"type": "string"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/Question"
						},
						"type": "array"
					}
				},
				"required": [
					"gameMode"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_HostInfo.personalResult-or-userInfo_": {
				"$ref": "#/components/schemas/Pick_HostInfo.Exclude_keyofHostInfo.personalResult-or-userInfo__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"EHostInfo": {
				"$ref": "#/components/schemas/Omit_HostInfo.personalResult-or-userInfo_"
			},
			"ResponseType__hostInfo-EHostInfo__": {
				"properties": {
					"data": {
						"properties": {
							"hostInfo": {
								"$ref": "#/components/schemas/EHostInfo"
							}
						},
						"required": [
							"hostInfo"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"Pick_GetHostInfoOpts.Exclude_keyofGetHostInfoOpts.personalResult-or-userInfo__": {
				"properties": {
					"activitiesBoard": {
						"type": "boolean"
					},
					"questions": {
						"type": "boolean"
					},
					"fullLeaderboard": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_GetHostInfoOpts.personalResult-or-userInfo_": {
				"$ref": "#/components/schemas/Pick_GetHostInfoOpts.Exclude_keyofGetHostInfoOpts.personalResult-or-userInfo__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"EGetHostInfo": {
				"$ref": "#/components/schemas/Omit_GetHostInfoOpts.personalResult-or-userInfo_"
			},
			"Game": {
				"properties": {
					"mode": {
						"$ref": "#/components/schemas/GameMode"
					},
					"name": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"mode",
					"name",
					"_id"
				],
				"type": "object"
			},
			"ResponseType__game-Game__": {
				"properties": {
					"data": {
						"properties": {
							"game": {
								"$ref": "#/components/schemas/Game"
							}
						},
						"required": [
							"game"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"ResponseType__game-Game--game_questions-Question-Array__": {
				"properties": {
					"data": {
						"properties": {
							"game_questions": {
								"items": {
									"$ref": "#/components/schemas/Question"
								},
								"type": "array"
							},
							"game": {
								"$ref": "#/components/schemas/Game"
							}
						},
						"required": [
							"game_questions",
							"game"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"ResponseType__list-Game-Array--total-number--page-number--limit-number--total_page-number__": {
				"properties": {
					"data": {
						"properties": {
							"total_page": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/Game"
								},
								"type": "array"
							}
						},
						"required": [
							"total_page",
							"limit",
							"page",
							"total",
							"list"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"GameListOpts": {
				"properties": {
					"limit": {
						"type": "integer",
						"format": "int32",
						"default": 10
					},
					"page": {
						"type": "integer",
						"format": "int32",
						"default": 1
					},
					"search": {
						"type": "string",
						"description": "Tìm kiếm"
					}
				},
				"type": "object"
			},
			"CmsUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"id"
				],
				"type": "object"
			},
			"ResponseType__user-CmsUser--token-any__": {
				"properties": {
					"data": {
						"properties": {
							"token": {},
							"user": {
								"$ref": "#/components/schemas/CmsUser"
							}
						},
						"required": [
							"token",
							"user"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"ResponseType__hostId-string__": {
				"properties": {
					"data": {
						"properties": {
							"hostId": {
								"type": "string"
							}
						},
						"required": [
							"hostId"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"PersonalResult": {
				"properties": {
					"accuracy": {
						"properties": {
							"percent": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"corrects": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"percent",
							"total",
							"corrects"
						],
						"type": "object"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"accuracy",
					"rank",
					"score"
				],
				"type": "object"
			},
			"Player": {
				"properties": {
					"score": {
						"type": "number",
						"format": "double"
					},
					"socketId": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"socketId",
					"avatar",
					"username",
					"id"
				],
				"type": "object"
			},
			"HostInfo": {
				"properties": {
					"questions": {
						"items": {
							"$ref": "#/components/schemas/Question"
						},
						"type": "array"
					},
					"groupId": {
						"type": "string"
					},
					"userInfo": {
						"$ref": "#/components/schemas/Player"
					},
					"activitiesBoard": {
						"items": {
							"$ref": "#/components/schemas/ActivityBoardItem"
						},
						"type": "array",
						"description": "Lịch sử hoạt động"
					},
					"personalResult": {
						"$ref": "#/components/schemas/PersonalResult"
					},
					"finalStandings": {
						"items": {
							"$ref": "#/components/schemas/HostLeaderboardItem"
						},
						"type": "array",
						"description": "Bảng xếp hạng"
					},
					"gameMode": {
						"$ref": "#/components/schemas/GameMode"
					},
					"state": {
						"$ref": "#/components/schemas/HostState"
					},
					"gameSettings": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"description": "Cài đặt game"
					},
					"gameId": {
						"type": "string"
					},
					"hostId": {
						"type": "string"
					}
				},
				"required": [
					"gameMode"
				],
				"type": "object"
			},
			"CreateHostBody": {
				"properties": {
					"hostInfo": {
						"$ref": "#/components/schemas/HostInfo"
					}
				},
				"required": [
					"hostInfo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseType__hostInfo-HostInfo__": {
				"properties": {
					"data": {
						"properties": {
							"hostInfo": {
								"$ref": "#/components/schemas/HostInfo"
							}
						},
						"required": [
							"hostInfo"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"GetHostInfoOpts": {
				"properties": {
					"questions": {
						"type": "boolean"
					},
					"userInfo": {
						"type": "boolean"
					},
					"activitiesBoard": {
						"type": "boolean"
					},
					"personalResult": {
						"type": "boolean"
					},
					"fullLeaderboard": {
						"type": "boolean"
					}
				},
				"type": "object"
			},
			"GetHostInfoBody": {
				"properties": {
					"hostId": {
						"type": "string"
					},
					"options": {
						"$ref": "#/components/schemas/GetHostInfoOpts"
					}
				},
				"required": [
					"hostId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseType__token-string__": {
				"properties": {
					"data": {
						"properties": {
							"token": {
								"type": "string"
							}
						},
						"required": [
							"token"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"GenTokenBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"hostId": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"hostId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseType__gameData-any__": {
				"properties": {
					"data": {
						"properties": {
							"gameData": {}
						},
						"required": [
							"gameData"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"ResponseType__players-Player-Array__": {
				"properties": {
					"data": {
						"properties": {
							"players": {
								"items": {
									"$ref": "#/components/schemas/Player"
								},
								"type": "array"
							}
						},
						"required": [
							"players"
						],
						"type": "object"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"code",
					"success"
				],
				"type": "object"
			},
			"GetPlayersBody": {
				"properties": {
					"hostId": {
						"type": "string"
					}
				},
				"required": [
					"hostId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"apiKey": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/external/host/create": {
			"post": {
				"operationId": "CreateHosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__hostIds-string-Array__"
								}
							}
						}
					}
				},
				"description": "Tạo hosts hàng loạt",
				"tags": [
					"External API"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ECreateHostBody"
							}
						}
					}
				}
			}
		},
		"/external/host/start": {
			"post": {
				"operationId": "StartHosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__hostIds-string-Array__"
								}
							}
						}
					}
				},
				"description": "Bắt đầu hosts hàng loạt",
				"tags": [
					"External API"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EHostIds"
							}
						}
					}
				}
			}
		},
		"/external/host/end": {
			"post": {
				"operationId": "StopHosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__hostIds-string-Array__"
								}
							}
						}
					}
				},
				"description": "Kết thúc hosts hàng loạt",
				"tags": [
					"External API"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EHostIds"
							}
						}
					}
				}
			}
		},
		"/external/host/info/{hostId}": {
			"post": {
				"operationId": "GetHostInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__hostInfo-EHostInfo__"
								}
							}
						}
					}
				},
				"description": "Lấy thông tin host",
				"tags": [
					"External API"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "hostId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EGetHostInfo"
							}
						}
					}
				}
			}
		},
		"/cms/game": {
			"post": {
				"operationId": "CreateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__game-Game__"
								}
							}
						}
					}
				},
				"description": "Tạo game mới",
				"tags": [
					"CMS API"
				],
				"security": [
					{
						"cmsBearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"game_questions": {
										"type": "string",
										"format": "binary"
									},
									"game_name": {
										"type": "string"
									},
									"game_mode": {
										"type": "string"
									}
								},
								"required": [
									"game_questions",
									"game_name",
									"game_mode"
								]
							}
						}
					}
				}
			}
		},
		"/cms/game/{game_id}": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__game-Game--game_questions-Question-Array__"
								}
							}
						}
					}
				},
				"description": "Lấy thông tin game",
				"tags": [
					"CMS API"
				],
				"security": [
					{
						"cmsBearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "game_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cms/game/update": {
			"post": {
				"operationId": "UpdateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__game-Game--game_questions-Question-Array__"
								}
							}
						}
					}
				},
				"description": "Cập nhật game",
				"tags": [
					"CMS API"
				],
				"security": [
					{
						"cmsBearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"game_questions": {
										"type": "string",
										"format": "binary"
									},
									"game_id": {
										"type": "string"
									},
									"game_name": {
										"type": "string"
									},
									"game_mode": {
										"type": "string"
									}
								},
								"required": [
									"game_questions",
									"game_id",
									"game_name",
									"game_mode"
								]
							}
						}
					}
				}
			}
		},
		"/cms/game/list": {
			"post": {
				"operationId": "ListGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__list-Game-Array--total-number--page-number--limit-number--total_page-number__"
								}
							}
						}
					}
				},
				"description": "Danh sách game",
				"tags": [
					"CMS API"
				],
				"security": [
					{
						"cmsBearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameListOpts"
							}
						}
					}
				}
			}
		},
		"/cms/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseType__user-CmsUser--token-any__"
								}
							}
						}
					}
				},
				"description": "Đăng nhập",
				"tags": [
					"CMS API"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}